Se actualiza la lógica de validación en el componente de pagos de clientes para verificar el campo 'total' en lugar de 'monto'. Se añaden nuevas funcionalidades para obtener estados y sucursales, y se ajusta la lógica de filtrado en la tabla de clientes. Además, se corrige la URL en el servicio de sucursales y se actualiza la versión en el pie de página.

+++ src/app/domains/ModAdmon/components/customers/customers.component.ts
@@ -3,7 +3,7 @@ import { RouterModule, ActivatedRoute } from '@angular/router';
 import { DomainsModule } from 'app/domains/domainsmodule';
 import { CellDoubleClickedEvent, ColDef, GridApi, GridReadyEvent, ICellRendererParams } from 'ag-grid-enterprise';
 import { alerts } from '../../../../helpers/alerts';
-
+import { States } from 'app/interface/states';
 import { catchError, concat, EMPTY, lastValueFrom, toArray } from 'rxjs';
 import { AgGridModule } from 'ag-grid-angular';
 import { ModalService } from 'app/services/modal.service';
@@ -13,9 +13,10 @@ import { CustomersPaymentsComponent } from './customers-payments.component';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { RadiusinfluenceComponent } from '../radiusinfluence/radiusinfluence.component';
 import { CustomersService } from 'app/services/customers.service';
-
 import { AutocompleteEditorComponent } from 'app/shared/autocomplete-editor/autocomplete-editor.component';
 import { InegiService } from 'app/services/inegi.service';
+import { BranchsService } from 'app/services/branchs.service';
+import { AuthService } from 'app/services/auth.service';
 
 @Component({
   selector: 'app-customers',
@@ -33,6 +34,8 @@ private signalsService     = inject(SignalsService);
 private modalService       = inject(NgbModal);
 private route              = inject(ActivatedRoute);
 private inegiService       = inject(InegiService);
+private branchesService    = inject(BranchsService);
+private authService        = inject(AuthService);
 
   ngOnInit() {
     this.obtenerDatos();
@@ -41,8 +44,9 @@ private inegiService       = inject(InegiService);
     
     this.route.data.subscribe(data => {
       this.type = data['type']; // 'CUSTOMERS' o 'PROVIDERS'
-
       this.obtenerDatos(); // Llamar a la función para cargar datos
+      this.getStates(); // Llamar a la función para obtener los estados
+      this.obtenerBranchs();
     });
   }
 
@@ -58,8 +62,15 @@ private inegiService       = inject(InegiService);
     effect(() => {
       this.idBranch = this.signalsService.getBranchSelectedBySidebar()();
       this.obtenerDatos();
+      this.obtenerBranchs();
       this.signalsService.deleteClientData();
     });
+
+    effect(() => {
+      this.idRoot = this.signalsService.getRootSelectedBySidebar()();
+      this.obtenerDatos();
+      this.obtenerBranchs();
+  })
   }
 
   @HostListener('window:beforeunload', ['$event'])
@@ -76,7 +87,7 @@ private inegiService       = inject(InegiService);
   notSavedChanges: boolean = false;
   selectedRowData: any = null;
   isOpen: boolean = false;
-
+  branchs: any[] = [];
   
   // Agregar esta nueva variable para almacenar el ID de la última fila editada
   private lastEditedRowId: number | string | null = null;
@@ -93,6 +104,7 @@ private inegiService       = inject(InegiService);
   idEmployee: number;
   infoCp: any;
 
+  private estados: string[] = []; // Agregar esta variable para almacenar los estados
 
   public defaultColDef: ColDef = {
     sortable: true,
@@ -146,11 +158,38 @@ private inegiService       = inject(InegiService);
   
   get colMaster(): ColDef[] {
     return [
-      { field: 'id', headerName: 'Id', editable: false, width: 53, hide : false,
+      { field: 'id', headerName: 'Id', editable: false, width: 110, hide : false,
         filter: 'agNumberColumnFilter', // Filtro para números (si el ID es numérico)
         filterParams: {
               filterOptions: ['equals'], // Opciones de filtro
      },
+    },
+    {
+      field: 'idBranch',
+      headerName: 'Nombre sucursal *',
+      headerClass: 'required-header',
+      hide: this.authService.hasDetailedPermission('principal', 'see-all-branches') ||
+        this.signalsService.getemailChoose() === 'root@beapp.com.mx' ? false : true,
+      editable: true,
+      filter: true,
+      width: 170,
+      cellEditor: 'agSelectCellEditor',
+      cellEditorParams: (params) => {
+        // Ensure depto data is available when creating editor
+        return {
+          values: this.branchs ? this.branchs.map((item) => item.id) : []
+        };
+      },
+      valueFormatter: (params) => {
+        // Handle potential null values and properly format the displayed value
+        if (!params.value) return '';
+
+        const foundBranch = this.branchs
+          ? this.branchs.find((item) => item.id === params.value)
+          : null;
+
+        return foundBranch ? foundBranch.name : params.value;
+      },
     },
       {
         field: 'company', headerName: 'Compania', editable: false, 
@@ -255,7 +294,17 @@ private inegiService       = inject(InegiService);
           return params.value;
         }
       },
-      { field: 'state', headerName: 'Estado', editable: true, filter: true, width: 160 },
+      {
+        field: 'state',
+        headerName: 'Estado',
+        filter: true,
+        width: 160,
+        editable: true,
+        cellEditor: 'agSelectCellEditor',
+        cellEditorParams: {
+          values: this.estados, // Usar la lista de estados obtenida
+        },
+      },
       { field: 'city', headerName: 'Ciudad', editable: true, width: 120, filter: true },
       {
         field: 'neighborhood',
@@ -332,6 +381,17 @@ private inegiService       = inject(InegiService);
     });
   }
 
+  obtenerBranchs() {
+    // alert('this.branchs'+ this.idBranch)
+    this.branchesService.getBrancheswoa(this.idRoot).subscribe(
+      (data: any) => {
+        this.branchs = data;
+        console.log(data)
+      },
+      (error) => console.error('Error fetching data:', error)
+    );
+  }
+
   onSelectedRow(event: any) {
     this.id = event.data.id;
   }
@@ -356,7 +416,8 @@ private inegiService       = inject(InegiService);
     if (event.colDef.field === 'cp') {
       event.data.neighborhood = '';
 
-      this.getZipCodeData(event.newValue).then((data: any) => {
+      setTimeout(async () => {
+        const data = await this.getZipCodeData(event.newValue);
         if (data && data.length > 0) {
           const cpData = data[0];
           event.data.state = cpData.estado;
@@ -364,7 +425,7 @@ private inegiService       = inject(InegiService);
 
           this.gridApi.applyTransaction({ update: [event.data] });
         }
-      });
+      }, 500);
     }
   }
 
@@ -523,6 +584,17 @@ private inegiService       = inject(InegiService);
     }
 
     const selectedData = selectedNodes[0].data;
+
+    // Validar que el préstamo sea 0 o no exista
+    if (selectedData.total && selectedData.total !== 0) {
+      alerts.basicAlert(
+        'Error al eliminar',
+        'No se puede eliminar mientras tenga facturas activas',
+        'error'
+      );
+      return;
+    }
+
     const id = selectedData.id;
     selectedData.active = 0;
     this.customerService.deleteCustomer(id).pipe(
@@ -573,20 +645,18 @@ private inegiService       = inject(InegiService);
       const selectedRowData = event.data; // Obtener los datos de la fila seleccionada
       const selectedId = selectedRowData.id; // Obtener el ID del registro
     
-      // Filtrar el grid para mostrar solo el registro con el ID seleccionado
-      const filterModel = {
-        id: {
-          type: 'equals',
-          filter: selectedId,
-        },
-      };
+      // Filtrar el grid para mostrar solo el registro con el ID seleccionado solo si la columna es "total"
+      if (colId === 'total') {
+        const filterModel = {
+          id: {
+            type: 'equals',
+            filter: selectedId,
+          },
+        };
     
-      this.gridApi.setFilterModel(filterModel);
-      this.gridApi.onFilterChanged();
-    
-      // Mostrar el componente <app-employeesxloans>
-      if(colId === 'total') {
-        this.activateCreditsTab();
+        this.gridApi.setFilterModel(filterModel);
+        this.gridApi.onFilterChanged();
+        this.activateCreditsTab(); // Activar la pestaña de créditos si es necesario
       }
      
       console.log('Datos ShowCredits:', this.showCreditsTab);
@@ -616,6 +686,19 @@ private inegiService       = inject(InegiService);
     adjustGridSize() {
       this.gridHeight = '20vh'; // Adjust as needed
     }
+
+  // Agregar esta función para obtener los estados
+  getStates() {
+    this.inegiService.getEstados().subscribe({
+      next: (data: { datos: States[] }) => {
+        this.estados = data.datos.map((estado) => estado.nom_agee);
+        this.estados.unshift('Sin estado'); // Agregar opción "Sin estado"
+      },
+      error: (error) => {
+        console.error('Error fetching states', error);
+      },
+    });
+  }
 }
